+++ src/main/java/SimpleDictionaryPersister.java
@@ -57,7 +57,7 @@ public class SimpleDictionaryPersister implements DictionaryPersister {
 	}
 	
 	@Override
-	public long getFrequency(String wordform) {
+	public int getFrequency(String wordform) {
 		if (wordform == null)
 			throw new NullPointerException();
